+++ src/commands/build-preact-app.js
@@ -1,11 +1,10 @@
 import path from 'path'
 
 import glob from 'glob'
-import ora from 'ora'
+import runSeries from 'run-series'
 
 import {getDefaultHTMLConfig} from '../appConfig'
 import webpackBuild from '../webpackBuild'
-import {logBuildResults} from '../webpackUtils'
 import cleanApp from './clean-app'
 
 // Using a config function as webpackBuild() sets NODE_ENV to production if it
@@ -52,15 +51,8 @@ function buildConfig(args) {
 export default function buildPreactApp(args, cb) {
   let dist = args._[2] || 'dist'
 
-  cleanApp({_: ['clean-app', dist]})
-
-  let spinner = ora(`Building Preact app`).start()
-  webpackBuild(args, buildConfig, (err, stats) => {
-    if (err) {
-      spinner.fail()
-      return cb(err)
-    }
-    logBuildResults(stats, spinner)
-    cb()
-  })
+  runSeries([
+    (cb) => cleanApp({_: ['clean-app', dist]}, cb),
+    (cb) => webpackBuild('Preact app', args, buildConfig, cb),
+  ], cb)
 }