+++ src/commands/clean-module.js
@@ -1,15 +1,5 @@
-import exec from '../exec'
+import {clean} from '../utils'
 
-import ora from 'ora'
-
-export default function cleanModule(args) {
-  let spinner = ora('Cleaning module').start()
-  try {
-    exec('rimraf', ['coverage', 'es', 'lib', 'umd'])
-    spinner.succeed()
-  }
-  catch (err) {
-    spinner.fail()
-    throw err
-  }
+export default function cleanModule(args, cb) {
+  clean('module', ['coverage', 'es', 'lib', 'umd'], cb)
 }