+++ src/moduleBuild.js
@@ -3,7 +3,7 @@ import path from 'path'
 
 import glob from 'glob'
 import ora from 'ora'
-import temp from 'temp'
+import runSeries from 'run-series'
 import merge from 'webpack-merge'
 
 import cleanModule from './commands/clean-module'
@@ -27,9 +27,9 @@ const DEFAULT_BABEL_IGNORE_CONFIG = [
 ]
 
 /**
- * Runs Babel with generated config written to a temporary .babelrc.
+ * Run Babel with generated config written to a temporary .babelrc.
  */
-function runBabel({copyFiles, outDir, src}, buildBabelConfig, userBabelConfig) {
+function runBabel(name, {copyFiles, outDir, src}, buildBabelConfig, userBabelConfig, cb) {
   let babelConfig = createBabelConfig(buildBabelConfig, userBabelConfig)
   babelConfig.ignore = DEFAULT_BABEL_IGNORE_CONFIG
 
@@ -40,71 +40,28 @@ function runBabel({copyFiles, outDir, src}, buildBabelConfig, userBabelConfig) {
 
   debug('babel config: %s', deepToString(babelConfig))
 
-  fs.writeFileSync('.babelrc', JSON.stringify(babelConfig, null, 2))
+  let spinner = ora(`Creating ${name} build`).start()
   try {
+    fs.writeFileSync('.babelrc', JSON.stringify(babelConfig, null, 2))
     exec('babel', babelArgs)
+    spinner.succeed()
+    fs.unlink('.babelrc', () => {
+      cb()
+    })
   }
-  finally {
-    fs.unlinkSync('.babelrc')
+  catch (err) {
+    spinner.fail()
+    fs.unlink('.babelrc', () => {
+      cb(err)
+    })
   }
 }
 
-export default function moduleBuild(args, buildConfig = {}, cb) {
-  // XXX Babel doesn't support passing the path to a babelrc file any more
-  if (glob.sync('.babelrc').length > 0) {
-    throw new UserError(
-      'Unable to build the module as there is a .babelrc in your project',
-      'nwb needs to write a temporary .babelrc to configure the build',
-    )
-  }
-
-  cleanModule(args)
-
-  let src = path.resolve('src')
-  let userConfig = getUserConfig(args)
-  let copyFiles = !!args['copy-files']
-
-  let spinner = ora('Creating ES5 build').start()
-  runBabel(
-    {copyFiles, outDir: path.resolve('lib'), src},
-    merge(buildConfig.babel, buildConfig.babelDev || {}, {
-      // Don't force ES5 users of the ES5 build to eat a .require
-      plugins: [require.resolve('babel-plugin-add-module-exports')],
-      // Don't set the path to nwb's babel-runtime, as it will need to be a
-      // peerDependency of your module if you use transform-runtime's helpers
-      // option.
-      setRuntimePath: false,
-    }),
-    userConfig.babel,
-  )
-  spinner.succeed()
-
-  // The ES6 modules build is enabled by default, and must be explicitly
-  // disabled if you don't want it.
-  if (userConfig.npm.esModules !== false) {
-    spinner = ora('Creating ES6 modules build').start()
-    runBabel(
-      {copyFiles, outDir: path.resolve('es'), src},
-      merge(buildConfig.babel, buildConfig.babelDev || {}, {
-        // Don't transpile modules, for use by ES6 module bundlers
-        modules: false,
-        // Don't set the path to nwb's babel-runtime, as it will need to be a
-        // peerDependency of your module if you use transform-runtime's helpers
-        // option.
-        setRuntimePath: false,
-      }),
-      userConfig.babel,
-    )
-    spinner.succeed()
-  }
-
-  temp.cleanupSync()
-
-  if (!userConfig.npm.umd) {
-    return cb()
-  }
-
-  spinner = ora('Creating UMD builds').start()
+/**
+ * Create development and production UMD builds for <script> tag usage.
+ */
+function buildUMD(args, buildConfig, userConfig, cb) {
+  let spinner = ora('Creating UMD builds').start()
 
   let pkg = require(path.resolve('package.json'))
   let entry = path.resolve(args._[1] || 'src/index.js')
@@ -125,7 +82,7 @@ export default function moduleBuild(args, buildConfig = {}, cb) {
   }
 
   process.env.NODE_ENV = 'development'
-  webpackBuild(args, webpackBuildConfig, (err, stats1) => {
+  webpackBuild(null, args, webpackBuildConfig, (err, stats1) => {
     if (err) {
       spinner.fail()
       return cb(err)
@@ -134,7 +91,7 @@ export default function moduleBuild(args, buildConfig = {}, cb) {
     webpackBuildConfig.babel = merge(buildConfig.babel, buildConfig.babelProd || {})
     webpackBuildConfig.devtool = 'source-map'
     webpackBuildConfig.output.filename = `${pkg.name}.min.js`
-    webpackBuild(args, webpackBuildConfig, (err, stats2) => {
+    webpackBuild(null, args, webpackBuildConfig, (err, stats2) => {
       if (err) {
         spinner.fail()
         return cb(err)
@@ -145,4 +102,62 @@ export default function moduleBuild(args, buildConfig = {}, cb) {
       cb()
     })
   })
+}
+
+export default function moduleBuild(args, buildConfig = {}, cb) {
+  // XXX Babel doesn't support passing the path to a babelrc file any more
+  if (glob.sync('.babelrc').length > 0) {
+    throw new UserError(
+      'Unable to build the module as there is a .babelrc in your project',
+      'nwb needs to write a temporary .babelrc to configure the build',
+    )
+  }
+
+  let src = path.resolve('src')
+  let userConfig = getUserConfig(args)
+  let copyFiles = !!args['copy-files']
+
+  let tasks = [
+    (cb) => cleanModule(args, cb),
+    (cb) => runBabel(
+      'ES5',
+      {copyFiles, outDir: path.resolve('lib'), src},
+      merge(buildConfig.babel, buildConfig.babelDev || {}, {
+        // Don't force ES5 users of the ES5 build to eat a .require
+        plugins: [require.resolve('babel-plugin-add-module-exports')],
+        // Don't set the path to nwb's babel-runtime, as it will need to be a
+        // peerDependency of your module if you use transform-runtime's helpers
+        // option.
+        setRuntimePath: false,
+      }),
+      userConfig.babel,
+      cb
+    )
+  ]
+
+  // The ES6 modules build is enabled by default, and must be explicitly
+  // disabled if you don't want it.
+  if (userConfig.npm.esModules !== false) {
+    tasks.push((cb) => runBabel(
+      'ES6 modules',
+      {copyFiles, outDir: path.resolve('es'), src},
+      merge(buildConfig.babel, buildConfig.babelDev || {}, {
+        // Don't transpile modules, for use by ES6 module bundlers
+        modules: false,
+        // Don't set the path to nwb's babel-runtime, as it will need to be a
+        // peerDependency of your module if you use transform-runtime's helpers
+        // option.
+        setRuntimePath: false,
+      }),
+      userConfig.babel,
+      cb
+    ))
+  }
+
+  // The UMD build must be explicitly enabled
+  if (userConfig.npm.umd) {
+    tasks.push((cb) => buildUMD(args, buildConfig, userConfig, cb))
+  }
+
+  runSeries(tasks, cb)
 }