+++ src/webpackBuild.js
@@ -1,3 +1,4 @@
+import ora from 'ora'
 import webpack from 'webpack'
 
 import createWebpackConfig from './createWebpackConfig'
@@ -5,8 +6,13 @@ import debug from './debug'
 import getPluginConfig from './getPluginConfig'
 import getUserConfig from './getUserConfig'
 import {deepToString, defaultNodeEnv} from './utils'
+import {logBuildResults} from './webpackUtils'
 
-export default function webpackBuild(args, buildConfig = {}, cb) {
+/**
+ * If you pass a non-falsy name, this will handle spinner display and output
+ * logging itself, otherwise use the stats provided in the callback.
+ */
+export default function webpackBuild(name, args, buildConfig = {}, cb) {
   defaultNodeEnv('production')
 
   let userConfig = getUserConfig(args)
@@ -19,6 +25,21 @@ export default function webpackBuild(args, buildConfig = {}, cb) {
 
   debug('webpack config: %s', deepToString(webpackConfig))
 
+  let spinner
+  if (name) {
+    spinner = ora(`Building ${name}`).start()
+  }
   let compiler = webpack(webpackConfig)
-  compiler.run(cb)
+  compiler.run((err, stats) => {
+    if (err) {
+      if (spinner) {
+        spinner.fail()
+      }
+      return cb(err)
+    }
+    if (spinner) {
+      logBuildResults(stats, spinner)
+    }
+    cb(null, stats)
+  })
 }